fix missing eRouterMode init for RPi
This fix (or something like it) is required to ensure that the LAN
interface is reliably started when running on RPi. Without a fix, there
is a race between eRouterMode being initialised and LAN_start() being
called. With systemd 216 startup is slow and race is always won, however
with systemd 230 (ie OE 2.2) startup is much faster and the race is
often lost.

eRouterMode needs to be initialised to something - otherwise the
"(bridge_mode == 0 && eRouterMode != 0)" test at the start of LAN_start()
will fail and the LAN will not be started. However, the exact details of
how eRouterMode should be initialised are not yet clear (e.g. should we
call GWP_SysCfgGetInt("last_erouter_mode") or just set to a fixed value?
If we do call GWP_SysCfgGetInt("last_erouter_mode") then should we also
call validate_mode() etc?).

Or perhaps eRouterMode should not be tested at all in LAN_start() when
building for RPi?

Change-Id: I91365b7e3e6fa9ed37f772c9a38e86bae4f70002
Signed-off-by: Andre McCurdy <armccurdy@gmail.com>